init: fix docker container with TS
