+++ app/static_content/index.html
@@ -2,8 +2,10 @@
 
 <html>
 <head>
+    <link rel="stylesheet" href="index.css">
+    <title>Index</title>
 </head>
-<body>
-<p> Hello from surge </p>
-</body>
+    <body>
+        <p> Hello from surge </p>
+    </body>
 </html>
