--- index.html
@@ -1,15 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-<head>
-    <meta charset="UTF-8">
-    <title>Доллару конец</title>
-</head>
-<body>
-    <h1>Госдолг сша превысил 621 миллиардов долларов</h1>
-</body>
-<style>
-    h1 {
-        font-family: 'Comic Sans MS', fantasy;
-    }
-</style>
-</html>
