+++ package.json
@@ -4,12 +4,13 @@
   "description": "The best in class messenger",
   "main": "index.js",
   "scripts": {
-    "start": "npm run-script build && node dist/index.js",
+    "dev": "npm run-script build && npm run-script start",
+    "start": "node dist/index.js",
     "build": "npx tsc",
     "test": "echo \"Error: no test specified\" && exit 1",
     "lint": "npm run-script lint-ts && npm run-script lint-css",
-    "lint-ts": "eslint **/*.ts --fix",
-    "lint-css": "npx stylelint **/*.css --fix"
+    "lint-ts": "eslint app/**/*.ts --fix",
+    "lint-css": "npx stylelint app/**/*.css --fix"
   },
   "repository": {
     "type": "git",
