+++ tsconfig.json
@@ -15,7 +15,7 @@
     // "sourceMap": true,                           /* Generates corresponding '.map' file. */
     // "outFile": "./",                             /* Concatenate and emit output to single file. */
     "outDir": "dist",                               /* Redirect output structure to the directory. */
-    "rootDir": ".",                                 /* Specify the root directory of input files. Use to control the output directory structure with --outDir. */
+    "rootDir": "app",                                 /* Specify the root directory of input files. Use to control the output directory structure with --outDir. */
     // "composite": true,                           /* Enable project compilation */
     // "tsBuildInfoFile": "./",                     /* Specify file to store incremental compilation information */
     // "removeComments": true,                      /* Do not emit comments to output. */
